Make wrap as a decorator and introduce @autowrap
The existing function tfplot.wrap() can actually be turned into
a python decorator providing the same functionality. To this end,
we use biwrap to support two usage of tfplot.wrap():

- As a function: tfplot.wrap(plotter, option='..')
- As a decorator:
  ```python
  @tfplot.wrap
  def plotter(tensor_val, option='..'):
    # ...
    return fig
  ```

to create a factory of plotting Tensors.

In the same vein, a new decorator @tfplot.autowrap is added.
It provides with additional features such as automatic injecting
matplotlib Figure and Axes objects if specified and handling of
return values (e.g. return ax directly). This will make most of
use cases easier.